# docs/swagger.yaml
basePath: /
definitions:
  main.UserLogin:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  main.UserRegister:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: RESTful API 文档示例
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gin Web API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户登录信息
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/main.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数校验失败或用户名密码错误
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户注册信息
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/main.UserRegister'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数校验失败或用户名已存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 用户注册
      tags:
      - auth
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"